move Diffusion class
