+++ examples/cfd/darcy_physics_informed/darcy_physics_informed_deeponet.py
@@ -27,41 +27,11 @@ from physicsnemo.utils.logging import LaunchLogger
 from physicsnemo.utils.checkpoint import save_checkpoint
 from physicsnemo.models.fno import FNO
 from physicsnemo.models.mlp import FullyConnected
-from sympy import Function, Number, Symbol
-
-from physicsnemo.sym.eq.pde import PDE
 from physicsnemo.sym.eq.phy_informer import PhysicsInformer
-
 from omegaconf import DictConfig
 from torch.utils.data import DataLoader
 
-from utils import HDF5MapStyleDataset
-
-
-class Diffusion(PDE):
-    """Diffusion equation: ``dT/dt - div(D * grad(T)) = Q``.
-
-    Equivalent to ``physicsnemo-sym``'s ``Diffusion`` class for the 2-D,
-    steady-state case with variable diffusivity ``D`` as a SymPy Function.
-
-    Reference: https://en.wikipedia.org/wiki/Diffusion_equation
-    """
-
-    def __init__(self, T="T", D="D", Q=0, dim=2, time=False):
-        self.dim = dim
-        x, y = Symbol("x"), Symbol("y")
-        iv = {"x": x, "y": y}
-        T_var = Function(T)(*iv.values())
-        D_var = Function(D)(*iv.values()) if isinstance(D, str) else Number(D)
-        Q_var = Number(Q) if isinstance(Q, (int, float)) else Q
-        self.equations = {
-            f"diffusion_{T}": (
-                (T_var.diff(Symbol("t")) if time else 0)
-                - (D_var * T_var.diff(x)).diff(x)
-                - (D_var * T_var.diff(y)).diff(y)
-                - Q_var
-            ),
-        }
+from utils import Diffusion, HDF5MapStyleDataset
 
 
 def validation_step(graph, dataloader, epoch):
